Update with proper process handling (I think)

+++ streamer.py
@@ -18,7 +18,9 @@ class Streamer:
       self.daemon = daemon
     
     def ensure_valid_stream( self, ytdlp_pid = None, ffmpeg_pid = None):
-        if not process_active(ytdlp_pid):
+        if not ytdlp_pid and self.stream:
+            ytdlp_pid = self.stream.pid
+        if ytdlp_pid and not process_active(ytdlp_pid):
             ytdlp_pid = None
         ffmpeg_pid = get_pid_by_name_and_args('ffmpeg', args_substring='/' + self.name + '/')#, exe_path=os.getcwd() + '/venv/bin') 
         if ytdlp_pid:
@@ -26,6 +28,8 @@ class Streamer:
                 return True
             else:
                 time.sleep(10)
+                # If we have yt-dlp but not ffmpeg it probably hasn't started yet.
+                # Maybe we want to add a recursion count but it hasn't been an issue yet.
                 return self.ensure_valid_stream( self, ytdlp_pid=ytdlp_pid )
         else:
             if ffmpeg_pid:
@@ -37,34 +41,32 @@ class Streamer:
     def start(self):
         def stream_thread():
             process_args = self.daemon.config["youtube-dl_cmd"].split(" ") + ["https://chaturbate.com/{}/".format(self.name), "--config-location", self.daemon.config["youtube-dl_config"]] 
-            stream = subprocess.Popen( process_args, 0, stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL )
+            self.stream = subprocess.Popen( process_args, 0, stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL )
             # sleep (10)? and make sure process didn't immediately exit
             poll_wait_time = self.daemon.config["process_poll_wait_time"]
             if poll_wait_time is None:
                 poll_wait_time = 10
             time.sleep( poll_wait_time )
-            if stream.poll() is None:
-                self.daemon.logger.info("Stream for {} appears to be healthy - validating.".format(self.name))
-                if self.ensure_valid_stream( ytdlp_pid = stream.pid ):
+            if self.stream and self.stream.poll() is None:
+                # self.daemon.logger.info("Stream for {} appears to be healthy - validating.".format(self.name))
+                if self.ensure_valid_stream( ytdlp_pid = self.stream.pid ):
                     self.daemon.logger.info("Started to record {}.".format(self.name))
-                    self.stream = stream
-                    stream.wait()
-            if self.stream:
-                self.stream = None
-                self.daemon.logger.info("Stopped {}.".format(self.name))
-            self.started = False # We marked as started even if we then failed - unmark
+                    self.started = True
+                    self.stream.wait()
             return
-        self.started = True
         thread = threading.Thread(target=stream_thread, args=())
         thread.daemon = True
         thread.start()
         return thread
 
-    def stop( self ):
+    def stop( self, signal_child = True ):
         if self.stream:
-            self.daemon.logger.info("Signaling {} to stop.".format(self.name))
-            self.stream.send_signal(signal.SIGINT)
-            self.stream.wait()
+            if signal_child:
+                self.daemon.logger.info("Signaling {} to stop.".format(self.name))
+                self.stream.send_signal(signal.SIGINT)
+                self.stream.wait()
             self.stream = None
-            self.daemon.logger.info("Stopped {}.".format(self.name))
+            if self.started:
+                self.started = False
+                self.daemon.logger.info("Stopped {}.".format(self.name))
